SD3 full and lora finetuning scripts
