+++ examples/stable-diffusion/training/README.md
@@ -416,6 +416,62 @@ python ../text_to_image_generation.py \
     --bf16
 ```
 
+## Fine-Tuning for Stable Diffusion 3
+
+### Full Model Fine-Tuning with SD3 (Stable Diffusion 3 and 3.5)
+
+We can use the `dog` dataset for the following examples.
+
+To launch SD3 full model training on single gaudi card, use:
+```bash
+python train_text_to_image_sd3.py \
+    --pretrained_model_name_or_path="stabilityai/stable-diffusion-3-medium-diffusers" \
+    --dataset_name="dog" \
+    --instance_prompt="a photo of sks dog" \
+    --validation_prompt="a photo of sks dog in a bucket" \
+    --output_dir="dog_ft_sd3" \
+    --mixed_precision="bf16" \
+    --resolution=1024 \
+    --train_batch_size=1 \
+    --learning_rate=5e-4 \
+    --max_grad_norm=1 \
+    --report_to="tensorboard" \
+    --lr_scheduler="constant" \
+    --lr_warmup_steps=0 \
+    --max_train_steps=1500 \
+    --validation_epochs=50 \
+    --save_validation_images \
+    --use_hpu_graphs_for_inference \
+    --use_hpu_graphs_for_training \
+    --gaudi_config_name="Habana/stable-diffusion" \
+    --sdp_on_bf16 \
+    --bf16 \
+    --guidance_scale 7 \
+```
+You can run training on multiple HPUs by replacing `python train_text_to_image_sd3.py`
+with `python ../../gaudi_spawn.py --world_size <num-HPUs> train_text_to_image_sd3.py`.
+
+> [!NOTE]
+> To use MPI for multi-card training, add `--use_mpi` after `--world_size <num-HPUs>`.
+> To use DeepSpeed instead of MPI, replace `--use_mpi` with `--use_deepspeed`.
+> Fine-tuning the full SD3.5-Large model requires multiple GPUs
+
+After training completes, you could directly use `text_to_image_generation.py` sample for inference as follows:
+```bash
+python ../text_to_image_generation.py \
+    --model_name_or_path dog_ft_sd3 \
+    --prompts "A picture of a sks dog in a bucket" \
+    --num_images_per_prompt 5 \
+    --batch_size 1 \
+    --num_inference_steps 28 \
+    --image_save_dir /tmp/sd3_images \
+    --use_habana \
+    --use_hpu_graphs \
+    --gaudi_config Habana/stable-diffusion \
+    --sdp_on_bf16 \
+    --bf16
+```
+
 ### DreamBooth LoRA Fine-Tuning with SD3 (Stable Diffusion 3 and 3.5)
 
 We can use the same `dog` dataset for the following examples.
@@ -429,63 +485,21 @@ python train_dreambooth_lora_sd3.py \
     --validation_prompt="a photo of sks dog in a bucket" \
     --output_dir="dog_lora_sd3" \
     --mixed_precision="bf16" \
+    --rank=8 \
     --resolution=1024 \
     --train_batch_size=1 \
-    --learning_rate=1e-4 \
-    --max_grad_norm=1 \
+    --learning_rate=5e-4 \
+    --max_grad_norm=0.5 \
     --report_to="tensorboard" \
     --lr_scheduler="constant" \
     --lr_warmup_steps=0 \
-    --max_train_steps=500 \
-    --seed="0" \
+    --max_train_steps=1500 \
+    --validation_epochs=50 \
+    --save_validation_images \
     --use_hpu_graphs_for_inference \
     --use_hpu_graphs_for_training \
     --gaudi_config_name="Habana/stable-diffusion" \
-    --bf16
-```
-
-stable-diffusion/training/train_text_to_image_sd3.py \
-  --pretrained_model_name_or_path stabilityai/stable-diffusion-3-medium-diffusers \
-  --dataset_name lambdalabs/naruto-blip-captions \
-  --instance_prompt="a cute naruto creature" \
-  --resolution 512 \
-  --center_crop \
-  --random_flip \
-  --train_batch_size 4 \
-  --max_train_steps 100 \
-  --learning_rate 1e-05 \
-  --lr_scheduler constant \
-  --lr_warmup_steps 0 \
-  --output_dir sd3_model_output \
-  --gaudi_config_name Habana/stable-diffusion \
-  --dataloader_num_workers 8 \
-  --bf16 \
-  --use_hpu_graphs_for_training \
-  --use_hpu_graphs_for_inference \
-  --validation_prompt="a cute naruto creature" \
-  --validation_epochs 48 \
-  --checkpointing_steps 2500
-
-
-You can run inference on multiple HPUs by replacing `python train_dreambooth_lora_flux.py`
-with `python ../../gaudi_spawn.py --world_size <num-HPUs> train_dreambooth_lora_flux.py`.
-
-> [!NOTE]
-> To use MPI for multi-card training, add `--use_mpi` after `--world_size <num-HPUs>`.
-> To use DeepSpeed instead of MPI, replace `--use_mpi` with `--use_deepspeed`.
-
-After training completes, you could directly use `text_to_image_generation.py` sample for inference as follows:
-```bash
-python ../text_to_image_generation.py \
-    --model_name_or_path "black-forest-labs/FLUX.1-dev" \
-    --lora_id dog_lora_flux \
-    --prompts "A picture of a sks dog in a bucket" \
-    --num_images_per_prompt 5 \
-    --batch_size 1 \
-    --image_save_dir /tmp/flux_images \
-    --use_habana \
-    --use_hpu_graphs \
-    --gaudi_config Habana/stable-diffusion \
+    --bf16 \
     --sdp_on_bf16 \
-    --bf16
+    --guidance_scale 7 \
 ```
