+++ examples/stable-diffusion/training/run.sh
@@ -1,4 +1,3 @@
-#python \
 python ../../gaudi_spawn.py --world_size 2 \
 train_text_to_image_sd3.py \
     --pretrained_model_name_or_path="stabilityai/stable-diffusion-3-medium-diffusers" \
@@ -9,16 +8,17 @@ train_text_to_image_sd3.py \
     --mixed_precision="bf16" \
     --resolution=1024 \
     --train_batch_size=1 \
-    --learning_rate=1e-4 \
+    --learning_rate=5e-4 \
     --max_grad_norm=1 \
     --report_to="tensorboard" \
     --lr_scheduler="constant" \
     --lr_warmup_steps=0 \
     --max_train_steps=500 \
-    --validation_epochs=20 \
+    --validation_epochs=50 \
     --save_validation_images \
     --use_hpu_graphs_for_inference \
     --use_hpu_graphs_for_training \
     --gaudi_config_name="Habana/stable-diffusion" \
     --sdp_on_bf16 \
-    --bf16
+    --bf16 \
+    --guidance_scale 7 \
