+++ examples/stable-diffusion/training/run_lora.sh
@@ -1,26 +1,25 @@
-#python ../../gaudi_spawn.py --world_size 2 \
 python \
 train_dreambooth_lora_sd3.py \
-    --pretrained_model_name_or_path="stabilityai/stable-diffusion-3.5-large" \
+    --pretrained_model_name_or_path="stabilityai/stable-diffusion-3-medium-diffusers" \
     --dataset_name="dog" \
     --instance_prompt="a photo of sks dog" \
     --validation_prompt="a photo of sks dog in a bucket" \
     --output_dir="dog_lora_sd3" \
     --mixed_precision="bf16" \
-    --rank=4 \
+    --rank=8 \
     --resolution=1024 \
     --train_batch_size=1 \
-    --learning_rate=1e-4 \
-    --max_grad_norm=1 \
+    --learning_rate=5e-4 \
+    --max_grad_norm=0.5 \
     --report_to="tensorboard" \
     --lr_scheduler="constant" \
     --lr_warmup_steps=0 \
-    --max_train_steps=500 \
+    --max_train_steps=1500 \
     --validation_epochs=50 \
     --save_validation_images \
     --use_hpu_graphs_for_inference \
     --use_hpu_graphs_for_training \
     --gaudi_config_name="Habana/stable-diffusion" \
-    --bf16
-
-#    --pretrained_model_name_or_path="stabilityai/stable-diffusion-3-medium-diffusers" \
+    --bf16 \
+    --sdp_on_bf16 \
+    --guidance_scale 7 \
